# doc/api.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/artist": {
      "post": {
        "operationId": "ArtistController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data for creating a new artist",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Vasia",
                  "grammy": true
                },
                "$ref": "#/components/schemas/CreateArtistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "artist successfully created",
            "schema": {
              "example": {
                "id": "123e4567-e89b-12d3-a456-426614174000",
                "name": "Vasia",
                "grammy": true
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseArtistDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          }
        },
        "summary": "Create artist",
        "tags": [
          "artist"
        ]
      },
      "get": {
        "operationId": "ArtistController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "artist found",
            "schema": {
              "example": {
                "name": "Vasia",
                "grammy": true
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseArtistDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all artist",
        "tags": [
          "artist"
        ]
      }
    },
    "/artist/{id}": {
      "get": {
        "operationId": "ArtistController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID artist id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "artist found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Vasia",
                    "grammy": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "artist not found"
          }
        },
        "summary": "Get artist by id",
        "tags": [
          "artist"
        ]
      },
      "put": {
        "operationId": "ArtistController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID artist id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Data for updating an existing artist",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Vasia",
                  "grammy": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "artist successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseArtistDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "artist not found"
          }
        },
        "summary": "Update artist",
        "tags": [
          "artist"
        ]
      },
      "delete": {
        "operationId": "ArtistController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID artist id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "artist successfully deleted"
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "artist not found"
          }
        },
        "summary": "Delete artist by id",
        "tags": [
          "artist"
        ]
      }
    },
...
    "/album": {
      "post": {
        "operationId": "AlbumController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data for creating a new album",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Album name",
                  "year": 2023,
                  "artistId": "123e4567-e89b-12d3-a456-426614174000"
                },
                "$ref": "#/components/schemas/CreateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "album successfully created",
            "schema": {
              "example": {
                "id": "123e4567-e89b-12d3-a456-426614174000",
                "name": "Album name",
                "year": 2023,
                "artistId": "123e4567-e89b-12d3-a456-426614174000"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseAlbumDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          }
        },
        "summary": "Create album",
        "tags": [
          "album"
        ]
      },
      "get": {
        "operationId": "AlbumController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "album found",
            "schema": {
              "example": {
                "name": "Album name",
                "year": 2023,
                "artistId": "123e4567-e89b-12d3-a456-426614174000"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseAlbumDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all album",
        "tags": [
          "album"
        ]
      }
    },
    "/album/{id}": {
      "get": {
        "operationId": "AlbumController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID album id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "album found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Album name",
                    "year": 2023,
                    "artistId": "123e4567-e89b-12d3-a456-426614174000"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "album not found"
          }
        },
        "summary": "Get album by id",
        "tags": [
          "album"
        ]
      },
      "put": {
        "operationId": "AlbumController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID album id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Data for updating an existing album",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Album name",
                  "year": 2023,
                  "artistId": "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "album successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseAlbumDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "album not found"
          }
        },
        "summary": "Update album",
        "tags": [
          "album"
        ]
      },
      "delete": {
        "operationId": "AlbumController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID album id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "album successfully deleted"
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "album not found"
          }
        },
        "summary": "Delete album by id",
        "tags": [
          "album"
        ]
      }
    },
...
    "/favs/artist/{id}": {
      "post": {
        "operationId": "FavoritesController_addArtist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID artist id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Add artist to favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseArtistDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. id is invalid (not uuid)"
          },
          "422": {
            "description": "artist not found"
          }
        },
        "summary": "Add artist to favorites",
        "tags": [
          "favorites"
        ]
      },
      "delete": {
        "operationId": "FavoritesController_removeArtist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID artist id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Remove artist from favorites"
          },
          "400": {
            "description": "Bad request. id is invalid (not uuid)"
          },
          "422": {
            "description": "artist not found"
          }
        },
        "summary": "Remove artist from favorites",
        "tags": [
          "favorites"
        ]
      }
    },
    "/favs/album/{id}": {
      "post": {
        "operationId": "FavoritesController_addAlbums",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID album id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Add album to favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseAlbumDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. id is invalid (not uuid)"
          },
          "422": {
            "description": "album not found"
          }
        },
        "summary": "Add album to favorites",
        "tags": [
          "favorites"
        ]
      },
      "delete": {
        "operationId": "FavoritesController_removeAlbums",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID album id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Remove album from favorites"
          },
          "400": {
            "description": "Bad request. id is invalid (not uuid)"
          },
          "422": {
            "description": "album not found"
          }
        },
        "summary": "Remove album from favorites",
        "tags": [
          "favorites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Home Library Service",
    "description": "Home music library service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateArtistDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "artist123",
            "description": "artist name"
          },
          "grammy": {
            "type": "boolean",
            "example": "true",
            "description": "artist grammy"
          }
        },
        "required": [
          "name",
          "grammy"
        ]
      },
      "ResponseArtistDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "artist123",
            "description": "artist name"
          },
          "grammy": {
            "type": "boolean",
            "example": "true",
            "description": "artist grammy"
          },
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid",
            "description": "Unique artist ID"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateAlbumDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the album",
            "example": "Album name"
          },
          "year": {
            "type": "number",
            "description": "Year of the album",
            "example": 2023
          },
          "artistId": {
            "type": "string",
            "description": "ID of the artist associated with the album",
            "example": "123e4567-e89b-12d3-a456-426614174000 | null"
          }
        },
        "required": [
          "name",
          "year",
          "artistId"
        ]
      },
      "ResponseAlbumDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique album ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the album",
            "example": "Album name"
          },
          "year": {
            "type": "number",
            "description": "Year of the album",
            "example": 2023
          },
          "artistId": {
            "type": "string",
            "description": "ID of the artist associated with the album",
            "example": "123e4567-e89b-12d3-a456-426614174000 | null"
          }
        },
        "required": [
          "id",
          "name",
          "year",
          "artistId"
        ]
      },
...
    }
  }
}